Squashed Commit:
Fixed not handling texture binding which is not preserved over the lifetime of a display list call.

Fixed not saving GL texture state messing up Sprites rendered outside of the GUI. Fixed Desktop not initialising values in it's 2nd constructor.

Removed glPushAttrib in favor of renderer.SetTexture() which is much faster. Added background texture to test app to detect Sprite glitches easier in the future.